refactor(mcp): extract outputSchema registry to access-mcp-output-schemas.ts
Moves the 393-line TOOL_OUTPUT_SCHEMAS registry + type helpers +
applyToolOutputSchemas() from access-mcp.ts to a focused module.
access-mcp.ts drops from 3396 to 2983 LOC (under the 3000 ratchet
without a baseline bump). No ratchet update needed.

The generic applyToolOutputSchemas<T extends {name;outputSchema?}>
preserves the tool type through the mapping. Import added to the top
import block; orphaned registry banner removed.

Surgical baseline update: 18 access-mcp.test.ts entries shifted to
new line numbers (one-for-one, same TS codes, compiler order).
