+++ scripts/test-typecheck-baseline.txt
@@ -143,24 +143,24 @@ tests/access-idempotency.test.ts(225,7): TS2349
 tests/access-idempotency.test.ts(294,7): TS2349
 tests/access-idempotency.test.ts(302,7): TS2349
 tests/access-idempotency.test.ts(397,7): TS2349
-tests/access-mcp.test.ts(9,22): TS7031
-tests/access-mcp.test.ts(31,26): TS7031
-tests/access-mcp.test.ts(44,23): TS7006
-tests/access-mcp.test.ts(64,28): TS7006
-tests/access-mcp.test.ts(78,27): TS7031
-tests/access-mcp.test.ts(88,31): TS7006
-tests/access-mcp.test.ts(99,32): TS7031
-tests/access-mcp.test.ts(109,23): TS7006
-tests/access-mcp.test.ts(123,29): TS7031
-tests/access-mcp.test.ts(134,33): TS7031
-tests/access-mcp.test.ts(156,29): TS7031
-tests/access-mcp.test.ts(156,35): TS7031
-tests/access-mcp.test.ts(156,46): TS7031
-tests/access-mcp.test.ts(156,60): TS7031
-tests/access-mcp.test.ts(156,69): TS7031
-tests/access-mcp.test.ts(156,89): TS7031
-tests/access-mcp.test.ts(184,29): TS7031
-tests/access-mcp.test.ts(184,42): TS7031
+tests/access-mcp.test.ts(10,22): TS7031
+tests/access-mcp.test.ts(32,26): TS7031
+tests/access-mcp.test.ts(45,23): TS7006
+tests/access-mcp.test.ts(65,28): TS7006
+tests/access-mcp.test.ts(81,27): TS7031
+tests/access-mcp.test.ts(91,31): TS7006
+tests/access-mcp.test.ts(104,32): TS7031
+tests/access-mcp.test.ts(114,23): TS7006
+tests/access-mcp.test.ts(128,29): TS7031
+tests/access-mcp.test.ts(139,33): TS7031
+tests/access-mcp.test.ts(163,29): TS7031
+tests/access-mcp.test.ts(163,35): TS7031
+tests/access-mcp.test.ts(163,46): TS7031
+tests/access-mcp.test.ts(163,60): TS7031
+tests/access-mcp.test.ts(163,69): TS7031
+tests/access-mcp.test.ts(163,89): TS7031
+tests/access-mcp.test.ts(191,29): TS7031
+tests/access-mcp.test.ts(191,42): TS7031
 tests/access-service.test.ts(432,35): TS2339
 tests/access-service.test.ts(433,35): TS2339
 tests/access-service.test.ts(2081,5): TS2349
